Add new attribute 18 "Ethernet PHY Configuration"
The Ethernet PHY Configuration attribute is an array of 16 bit long
bitfields used to configure properties of the Ethernet PHY. Each
element of the array represents one Ethernet port of the module.
Bit 0 will be used to configure fallback duplex if the PHY fails to
auto negotiate the duplex. Default is half duplex but with this new
configuration option the module can be forced to use full duplex.

Issue m#0016355

Change-Id: I013e369931522b7a7a31835b0c1ee0e3245fe33c