Corrections for compiler specific struct pack macro
In abp.h there is an macro defined telling the compiler
to pack structures. This macros should never been defined in
abp.h in the first place, but since it's there and has been for
years it can't be removed. Also since the macro doesn't
cover all waus different compilers use to pack structures additional
macros are needed to cover those cases for the structures that need
to be packed. m#0017237.

Change-Id: Ibb378cdf5956989d57114766868ca75fae90851e