Removed the serial number attribute from the BACnet host object
The attribute was recently added and has not yet been included in a
released version of ABP and can therefore be removed safely. It was
decided by the product owner of ABCC40 BACnet/IP to initially not
support a BACnet specific serial number attribute. Instead the device
will only support the Anybus generic serial number attribute of the
Application host object.

Change-Id: Ib240b44703feb8752838d1b94454d483e360b1a3